requirements.txt
APScheduler==3.6.3
asgiref==3.2.7
bcrypt==3.1.7
cffi==1.14.0
cryptography==2.9.2
dj-database-url==0.5.0
Django==3.0.6
django-filter==2.2.0
django-heroku==0.3.1
djangorestframework==3.11.0
gunicorn==20.0.4
importlib-metadata==1.6.0
Markdown==3.2.2
paramiko==2.7.1
psycopg2==2.8.5
pycparser==2.20
PyNaCl==1.3.0
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
tzlocal==2.1
whitenoise==5.0.1
zipp==3.1.0
